# api/openapi.yaml
openapi: "3.0.3"
info:
  title: PingCast API
  version: "2.0.0"
  description: Uptime monitoring API with multi-type checks
paths:
  /api/auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/incidents:
    post:
      operationId: createIncident
      tags: [incidents]
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Manual incident opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "402":
          description: Pro subscription required
        "403":
          description: Not the owner of the monitor
        "422":
          description: Validation error
...
  /api/import/atlassian:
    post:
      operationId: importAtlassian
      tags: [import]
      security:
        - sessionAuth: []
        - bearerAuth: []
      description: |
        Accepts a raw Atlassian Statuspage v1 JSON export body. Creates
        monitors from components, incidents with preserved states, and
        the full incident_updates timeline. Pro-gated; all-or-nothing
        transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Import result counters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtlassianImportResult"
        "400":
          description: Invalid or unsupported export
        "402":
          description: Pro subscription required
  /api/custom-domains:
    get:
      operationId: listCustomDomains
      tags: [domains]
      security:
        - sessionAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: User's custom domains (all statuses)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomDomain"
    post:
      operationId: requestCustomDomain
      tags: [domains]
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname]
              properties:
                hostname: { type: string }
      responses:
        "201":
          description: Pending — set the CNAME + serve the .well-known probe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomDomain"
        "402": { description: Pro required }
        "422": { description: Invalid hostname }
...
  /api/maintenance-windows:
    get:
      operationId: listMaintenanceWindows
      tags: [maintenance]
      security:
        - sessionAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: User's maintenance windows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceWindow"
    post:
      operationId: scheduleMaintenanceWindow
      tags: [maintenance]
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleMaintenanceWindowRequest"
      responses:
        "201":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceWindow"
        "402":
          description: Pro required
...
  /api/me/branding:
    get:
      operationId: getMyBranding
      tags: [branding]
      security:
        - sessionAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Current branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branding"
    patch:
      operationId: updateMyBranding
      tags: [branding]
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branding"
      responses:
        "200":
          description: Updated branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branding"
        "402":
          description: Pro subscription required
...
  /api/public/lookup-domain:
    get:
      operationId: lookupCustomDomain
      tags: [domains]
      description: |
        Given a hostname, returns the status-page slug it resolves to.
        Consumed by the Next.js middleware to route custom-domain
        requests. Public, cached 5 min. 404 if the hostname isn't an
        active custom domain.
      parameters:
        - name: hostname
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Mapped slug
          content:
            application/json:
              schema:
                type: object
                required: [slug]
                properties:
                  slug: { type: string }
        "404":
          description: Hostname not registered or not active
...
  /api/newsletter/subscribe:
    post:
      operationId: subscribeToNewsletter
      tags: [newsletter]
      description: Public. Signs a visitor up to the global PingCast newsletter (blog + product updates). Double-opt-in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                source: { type: string, description: "Optional tag for where the signup happened: footer, blog_sidebar, post_cta:<slug>" }
      responses:
        "202": { description: Acknowledged — check your inbox to confirm. }
        "400": { description: Invalid email }
  /api/newsletter/confirm:
    get:
      operationId: confirmNewsletterSubscription
      tags: [newsletter]
      description: Public. Clicked from the confirmation email. Flips confirmed_at.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Confirmation page (text/html) }
        "404": { description: Invalid or expired token }
  /api/newsletter/unsubscribe:
    get:
      operationId: unsubscribeFromNewsletter
      tags: [newsletter]
      description: Public. Clicked from any outbound email. Deletes the subscriber.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Unsubscribed (text/html) }
        "404": { description: Invalid token }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key with pc_live_ prefix"
  schemas:
    RegisterRequest:
      type: object
      required: [email, slug, password]
      properties:
        email:
          type: string
          format: email
        slug:
          type: string
          pattern: "^[a-z0-9-]{3,30}$"
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        session_id:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [free, pro]
        created_at:
          type: string
          format: date-time
...
    Incident:
      type: object
      required: [id, monitor_id, started_at, cause, state, is_manual]
      properties:
        id:
          type: integer
          format: int64
        monitor_id:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        cause:
          type: string
        state:
          type: string
          enum: [investigating, identified, monitoring, resolved]
        is_manual:
          type: boolean
        title:
          type: string
          nullable: true
    CreateIncidentRequest:
      type: object
      required: [monitor_id, title, body]
      properties:
        monitor_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 3
          maxLength: 200
        body:
          type: string
          minLength: 1
          maxLength: 4000
...
    AtlassianImportResult:
      type: object
      required: [monitors_created, incidents_created, updates_created, components_skipped]
      properties:
        monitors_created: { type: integer }
        incidents_created: { type: integer }
        updates_created: { type: integer }
        components_skipped: { type: integer }
    CustomDomain:
      type: object
      required: [id, hostname, validation_token, status, created_at]
      properties:
        id:               { type: integer, format: int64 }
        hostname:         { type: string }
        validation_token: { type: string, description: "Serve this token at /.well-known/pingcast/<token>" }
        status:           { type: string, enum: [pending, validated, active, failed] }
        last_error:       { type: string, nullable: true }
        dns_validated_at: { type: string, format: date-time, nullable: true }
        cert_issued_at:   { type: string, format: date-time, nullable: true }
        created_at:       { type: string, format: date-time }
...
    MaintenanceWindow:
      type: object
      required: [id, monitor_id, starts_at, ends_at, reason, created_at]
      properties:
        id:         { type: integer, format: int64 }
        monitor_id: { type: string,  format: uuid }
        starts_at:  { type: string,  format: date-time }
        ends_at:    { type: string,  format: date-time }
        reason:     { type: string }
        created_at: { type: string,  format: date-time }
    ScheduleMaintenanceWindowRequest:
      type: object
      required: [monitor_id, starts_at, ends_at, reason]
      properties:
        monitor_id: { type: string,  format: uuid }
        starts_at:  { type: string,  format: date-time }
        ends_at:    { type: string,  format: date-time }
        reason:     { type: string,  minLength: 1, maxLength: 500 }
...
    Branding:
      type: object
      properties:
        logo_url:
          type: string
          nullable: true
        accent_color:
          type: string
          nullable: true
          description: "Hex like #3b82f6. Rendered as --accent CSS var on the status page."
        custom_footer_text:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string